changed float to double + minor C++ improvement, it failed compiling on my Mac otherwise (ASKAPSDP-1672)

+++ synthesis/current/gridding/SphFuncVisGridder.h
@@ -122,7 +122,7 @@ namespace askap
 			private:
 			
 				/// @brief prolate spheroidal alpha parameter
-                float itsAlpha;
+                double itsAlpha;
 
 		};
